EDIcraft-main/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents
rcon
osdu-client
pytest
boto3
